backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow
requests==2.31.0
python-dotenv==1.0.0
replicate==0.25.1
anyio>=4.0.0